+++ .gitignore
@@ -14,3 +14,4 @@ node_modules
 .npm
 .node_repl_history
 coverage
+dist
