+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "farmbotjs",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Javascript wrapper for Farmbot's authentication and RPC instructions",
-  "main": "src/index.js",
+  "main": "dist/farmbot.js",
   "scripts": {
     "test": "karma start karma.conf.js"
   },
@@ -27,6 +27,7 @@
     "karma-chrome-launcher": "^0.2.1",
     "karma-coverage": "^0.5.3",
     "karma-jasmine": "^0.3.6",
-    "karma-spec-reporter": "0.0.23"
+    "karma-spec-reporter": "0.0.23",
+    "source-map": "^0.5.3"
   }
 }
